Minor fixes

+++ tinysys/SDK/usbc.h
@@ -51,7 +51,7 @@ extern volatile uint32_t *IO_USBCSTA;
 
 struct USBCDCLineCoding
 {
-    uint16_t dwDTERate;
+    uint32_t dwDTERate;
 	uint8_t bCharFormat;
 	uint8_t bParityType;
 	uint8_t bDataBits;
